Move React components into component folder
