+++ src/prev/index.js
@@ -1,23 +1,29 @@
 // All exposed UI components
-export { default as Text } from './Text';
-export { default as Heading } from './Heading';
-export { default as TextLink } from './Button/TextLink';
-export { default as LinkButton } from './Button/LinkButton';
-export { default as ActionButton } from './Button/ActionButton';
-export { default as Icon } from './Icon';
-export { default as Image } from './Image';
-export { default as ItemBox } from './ItemBox';
-export { default as Sidebar } from './Sidebar';
-export { default as Grid } from './Grid';
-export { default as FieldSelect } from './Form/Fields/FieldSelect';
-export { default as FieldText } from './Form/Fields/FieldText';
-export { default as FieldRadio } from './Form/Fields/FieldRadio';
+export { default as Text } from './components/Text';
+export { default as Heading } from './components/Heading';
+export { default as TextLink } from './components/Button/TextLink';
+export { default as LinkButton } from './components/Button/LinkButton';
+export { default as ActionButton } from './components/Button/ActionButton';
+export { default as Icon } from './components/Icon';
+export { default as Image } from './components/Image';
+export { default as ItemBox } from './components/ItemBox';
+export { default as Sidebar } from './components/Sidebar';
+export { default as Grid } from './components/Grid';
+export { default as FieldSelect } from './components/Form/Fields/FieldSelect';
+export { default as FieldText } from './components/Form/Fields/FieldText';
+export { default as FieldRadio } from './components/Form/Fields/FieldRadio';
 export {
   default as FieldDollarHandles,
-} from './Form/Fields/FieldDollarHandles';
-export { default as FieldCheckbox } from './Form/Fields/FieldCheckbox';
-export { default as FieldTextArea } from './Form/Fields/FieldTextArea';
-export { default as FieldFileUpload } from './Form/Fields/FieldFileUpload';
-export { default as FormField } from './Form/FormField';
-export { default as FormHeader } from './Form/FormHeader';
-export { default as FormNavigation } from './Form/FormNavigation';
+} from './components/Form/Fields/FieldDollarHandles';
+export {
+  default as FieldCheckbox,
+} from './components/Form/Fields/FieldCheckbox';
+export {
+  default as FieldTextArea,
+} from './components/Form/Fields/FieldTextArea';
+export {
+  default as FieldFileUpload,
+} from './components/Form/Fields/FieldFileUpload';
+export { default as FormField } from './components/Form/FormField';
+export { default as FormHeader } from './components/Form/FormHeader';
+export { default as FormNavigation } from './components/Form/FormNavigation';
