+++ stories/index.js
@@ -1,13 +1,13 @@
 import { storiesOf } from '@storybook/react';
 
 // Stories.
-import '../src/prev/Text/stories';
-import '../src/prev/Button/stories';
-import '../src/prev/Icon/stories';
-import '../src/prev/Form/Fields/stories';
-import '../src/prev/Form/FormGrid/stories';
-import '../src/prev/ItemBox/stories';
-import '../src/prev/Grid/stories';
+import '../src/prev/components/Text/stories';
+import '../src/prev/components/Button/stories';
+import '../src/prev/components/Icon/stories';
+import '../src/prev/components/Form/Fields/stories';
+import '../src/prev/components/Form/FormGrid/stories';
+import '../src/prev/components/ItemBox/stories';
+import '../src/prev/components/Grid/stories';
 import '../src/components/Doughnut/stories';
 
 storiesOf('Information', module).add('Current versions', () => (
